Visualizao dos trabalhos submetidos para o Admin_fecitec
