+++ admin_fecitec/templates/admin_screen_submission.html
@@ -29,79 +29,63 @@
     <div class="table-container">
         <table class="responsive-table">
             <thead>
+                
                 <tr>
                     <th></th>
                     <th scope="col">Participante</th>
                     <th scope="col">Título</th>
                     <th scope="col">Data de Submissão</th>
                     <th scope="col">Situação</th>
+                    <th scope="col">Responsaveis Atuais <th/>
                     <th scope="col"></th>
                 </tr>
+                
             </thead>
             <tbody>
-              
-                <tr>
-                    <td>
-                        <div class="participant_phonto"> <img class="participant_phonto" src="https://avatars.githubusercontent.com/u/148590574?v=4" alt=""></div>
-                    </td>
-                    <td>1</td>
-                    <td>2</td>
-                    <td class="container-status">
-                        3
-                    </td>
-                    <td>
-                        4
-                    </td>
-                    <td>
-                        <div class="dropdown">
-                            <span class="dropdown-button">...</span>
-                            <div class="dropdown-content">
-                                <a style="color:#79B6DE;" href="#" class="editar">
-                                    Editar
-                                    <i class="fa-solid fa-pen-to-square"></i>
-                                </a>
-                                <a style="color: red;" href="#" class="cancelar">
-                                    Cancelar
-                                    <i class="fa-solid fa-ban"></i>
-                                </a>
+                {% for  submission in submissions %}
+                    <tr>
+                        <td>
+                            {{submission.id}}
+                            {% comment %} <div class="participant_phonto"> <img class="participant_phonto" src="https://avatars.githubusercontent.com/u/148590574?v=4" alt=""></div> {% endcomment %}
+                        </td>
+                        <td>
+                            {{submission.participante.nome}}
+                        </td>
+                        <td>
+                            titulo nao pego
+                        </td>
+                        <td>
+                            {{submission.submission_date}}
+                        </td>
+                        <td class="container-status">
+                            {{submission.status}}
+                        </td>
+                        <td>
+                            Respnosavais atuais(comissao)
+                        </td>
+                        <td>
+                            <div class="dropdown">
+                                <span class="dropdown-button">...</span>
+                                <div class="dropdown-content">
+                                    <a style="color:#79B6DE;" href="#" class="editar">
+                                        Editar
+                                        <i class="fa-solid fa-pen-to-square"></i>
+                                    </a>
+                                    <a style="color: red;" href="#" class="cancelar">
+                                        Cancelar
+                                        <i class="fa-solid fa-ban"></i>
+                                    </a>
+                                </div>
                             </div>
-                        </div>
-                    </td>
-                </tr>
+                        </td>
+                    </tr>
 
-                <tr>
-                    <td>
-                        <div class="participant_phonto"> <img class="participant_phonto" src="https://avatars.githubusercontent.com/u/148590574?v=4" alt=""></div>
-                    </td>
-                    <td>1</td>
-                    <td>2</td>
-                    <td class="container-status">
-                        3
-                    </td>
-                    <td>
-                        4
-                    </td>
-                    <td>
-                        <div class="dropdown">
-                            <span class="dropdown-button">...</span>
-                            <div class="dropdown-content">
-                                <a style="color:#79B6DE;" href="#" class="editar">
-                                    Editar
-                                    <i class="fa-solid fa-pen-to-square"></i>
-                                </a>
-                                <a style="color: red;" href="#" class="cancelar">
-                                    Cancelar
-                                    <i class="fa-solid fa-ban"></i>
-                                </a>
-                            </div>
-                        </div>
-                    </td>
-                </tr>
-        
+                    
+                {% endfor %}
             </tbody>
         </table>
       </div>
-    </div> <!--- fim do content father dentro do main base --->
+    </div> <!--- fim do content father dentro do main base -->
