+++ fecitec/forms.py
@@ -2,7 +2,7 @@ from django import forms
 from django.core.mail.message import EmailMessage
 import re
 from django.core.exceptions import ValidationError
-from .models import SubmissionToWork
+
 
 def validate_phone_number(value):
     phone_pattern = re.compile(r'^\+?1?\d{9,15}$')
@@ -41,8 +41,3 @@ class ContactForm(forms.Form):
         )
 
         mail.send()
-
-class SubmissionToWorkForm(forms.ModelForm):
-    class Meta:
-        model = SubmissionToWork
-        fields = ['school_name', 'area', 'municipality', 'title', 'sub_area', 'summary', 'state', 'formation', 'arquivo_modelo', 'form_of_presentation', 'participante']
